backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
websockets==14.1
pyyaml==6.0.2
Pillow==11.0.0
python-multipart==0.0.18
pydantic==2.10.0
pydantic-settings==2.6.0
